refactor: response를 별도 패키지로 분리하여 가독성 증진
