+++ src/main/java/com/techfork/post/presentation/PostControllerV2.java
@@ -1,7 +1,7 @@
 package com.techfork.post.presentation;
 
-import com.techfork.post.presentation.CompanyListResponse;
-import com.techfork.post.presentation.PostListResponse;
+import com.techfork.post.presentation.response.CompanyListResponse;
+import com.techfork.post.presentation.response.PostListResponse;
 import com.techfork.post.application.query.result.GetCompanyListResult;
 import com.techfork.post.application.query.result.GetPostListResult;
 import com.techfork.post.application.query.GetPostsByCompanyV2Query;
